# Improve listpack Handling and Decoding Efficiency: 16.3% improvement on LRANGE command