--- a/src/listpack.c
+++ b/src/listpack.c
@@ -676,8 +676,7 @@ static inline unsigned char *lpGetWithSize(unsigned char *p, int64_t *count, uns
 /* Return the listpack element pointed by 'p'.
  *
  * The function has the same behaviour as lpGetWithSize when 'entry_size' is NULL,
- * but avoids a lot of unecesarry branching performance penalties.
- * */
+ * but avoids a lot of unecesarry branching performance penalties. */
 static inline unsigned char *lpGetWithBuf(unsigned char *p, int64_t *count, unsigned char *intbuf) {
     int64_t val;
     uint64_t uval, negstart, negmax;