--- a/src/listpack.c
+++ b/src/listpack.c
@@ -683,24 +683,13 @@ static inline unsigned char *lpGetWithBuf(unsigned char *p, int64_t *count, unsi
     const unsigned char encoding = p[0];
 
     assert(p); /* assertion for valgrind (avoid NPD) */
-    /* string encoding */
-    if (LP_ENCODING_IS_6BIT_STR(encoding)) {
-        *count = LP_ENCODING_6BIT_STR_LEN(p);
-        return p+1;
-    }
-    if (LP_ENCODING_IS_12BIT_STR(encoding)) {
-        *count = LP_ENCODING_12BIT_STR_LEN(p);
-        return p+2;
-    }
-    if (LP_ENCODING_IS_32BIT_STR(encoding)) {
-        *count = LP_ENCODING_32BIT_STR_LEN(p);
-        return p+5;
-    }
-    /* int encoding */
     if (LP_ENCODING_IS_7BIT_UINT(encoding)) {
         negstart = UINT64_MAX; /* 7 bit ints are always positive. */
         negmax = 0;
         uval = encoding & 0x7f;
+    } else if (LP_ENCODING_IS_6BIT_STR(encoding)) {
+        *count = LP_ENCODING_6BIT_STR_LEN(p);
+        return p+1;
     } else if (LP_ENCODING_IS_13BIT_INT(encoding)) {
         uval = ((encoding&0x1f)<<8) | p[1];
         negstart = (uint64_t)1<<12;
@@ -734,6 +723,13 @@ static inline unsigned char *lpGetWithBuf(unsigned char *p, int64_t *count, unsi
                (uint64_t)p[8]<<56;
         negstart = (uint64_t)1<<63;
         negmax = UINT64_MAX;
+    } else if (LP_ENCODING_IS_12BIT_STR(encoding)) {
+        *count = LP_ENCODING_12BIT_STR_LEN(p);
+        return p+2;
+    }
+    if (LP_ENCODING_IS_32BIT_STR(encoding)) {
+        *count = LP_ENCODING_32BIT_STR_LEN(p);
+        return p+5;
     } else {
         uval = 12345678900000000ULL + encoding;
         negstart = UINT64_MAX;