This PR focused on refining listpack encoding/decoding functions and optimizing reply handling mechanisms related to it. 
Each commit has the measured improvement up until the last accumulated improvement of 16.3% on [memtier_benchmark-1key-list-100-elements-lrange-all-elements-pipeline-10](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-1key-list-100-elements-lrange-all-elements-pipeline-10.yml) benchmark. 

Connection mode | CE Baseline  (Nov 14th) 701f06657d20ec42f0cf78de0ac9d7197e44c00c | CE PR #13652 | CE PR vs CE Unstable
-- | -- | -- | --
TCP | 155696 | 178874 | 14.9%
Unix socket | 169743 | 197428 | 16.3%






To test it we can simply focus on the scan.tcl

```
tclsh tests/test_helper.tcl --single unit/replybufsize
```

### Commit details:
- 2e58d048fd9d1365c56292ebd69f2ed141bfeda1 + 29c6c86c6b96376f807656f203b61d0e7fcb65a4 : Eliminate an indirect memory access on lpCurrentEncodedSizeBytes and completely avoid passing p* fully to lpCurrentEncodedSizeBytes + Add lpNextWithBytes helper function and optimize addListListpackRangeReply 

    **- Improvement of 3.1%, from 168969.88 ops/sec to 174239.75 ops/sec**
- af52aacff86e2809f3451e2561043d532b9ee223 Refactor lpDecodeBacklen for loop-based decoding, improving readability and branch efficiency.
    **- NO CHANGE. REVERTED in 09f6680ba0d0b5acabca537c651008f0c8ec061b**
    
- 048bfe4edaaf5671f7f06b06d1492f81e6af3e59 + 03e8ff3af70891cbd3d53ca865558976459bb38e : reducing condition checks in _addReplyToBuffer, inlining it, and avoid entering it when there are there already entries in the reply list
 and check if the reply length exceeds available buffer space before calling _addReplyToBuffer 
    **- accumulated Improvement of 12.4%, from 168969.88 ops/sec to 189726.81 ops/sec**

- 9a63d4d6a9fa946505e31ecce4c7796845fc022c:  always update the buf_peak on _addReplyToBufferOrList
    **- accumulated Improvement of 14.2%, from 168969.88 ops/sec to 193887 ops/sec**

- b544ade67628a1feaf714d6cfd114930e0c7670b:  Introduce lpEncodeBacklenBytes to avoid any indirect memory access on previous usage of lpEncodeBacklen(NULL,...). inline lpEncodeBacklenBytes().
    **- accumulated Improvement of 16.3%, from 168969.88 ops/sec to 197427.70 ops/sec**

### To replicate:

Populate key with 100 entries lp encoded
```
redis-cli -p 6380 "LPUSH" "list:100" "vyoomgwuzv" "xamjodnbpf" "ewomnmugfa" "ljcgdooafo" "pcxdhdjwnf" "djetcyfxuc" "licotqplim" "alqlzsvuuz" "ijsmoyesvd" "whmotknaff" "rkaznetutk" "ksqpdywgdd" "gorgpnnqwr" "gekntrykfh" "rjkknoigmu" "luemuetmia" "gxephxbdru" "ncjfckgkcl" "hhjclfbbka" "cgoeihlnei" "zwnitejtpg" "upodnpqenn" "mibvtmqxcy" "htvbwmfyic" "rqvryfvlie" "nxcdcaqgit" "gfdqdrondm" "lysbgqqfqw" "nxzsnkmxvi" "nsxaigrnje" "cwaveajmcz" "xsepfhdizi" "owtkxlzaci" "agsdggdghc" "tcjvjofxtd" "kgqrovsxce" "ouuybhtvyb" "ueyrvldzwl" "vpbkvwgxsf" "pytrnqdhvs" "qbiwbqiubb" "ssjqrsluod" "urvgxwbiiz" "ujrxcmpvsq" "mtccjerdon" "xczfmrxrja" "imyizmhzjk" "oguwnmniig" "mxwgdcutnb" "pqyurbvifk" "ccagtnjilc" "mbxohpancs" "lgrkndhekf" "eqlgkwosie" "jxoxtnzujs" "lbtpbknelm" "ichqzmiyot" "mbgehjiauu" "aovfsvbwjg" "nmgxcctxpn" "vyqqkuszzh" "rojeolnopp" "ibhohmfxzt" "qbyhorvill" "nhfnbxqgol" "wkbasfyzqz" "mjjuylgssm" "imdqxmkzdj" "oapbvnisyq" "bqntlsaqjb" "ocrcszcznp" "hhniikmtsx" "hlpdstpvzw" "wqiwdbncmt" "vymjzlzqcn" "hhjchwjlmc" "ypfeltycpy" "qjyeqcfhjj" "uapsgmizgh" "owbbdezgxn" "qrosceblyo" "sahqeskveq" "dapacykoah" "wvcnqbvlnf" "perfwnpvkl" "ulbrotlhze" "fhuvzpxjbc" "holjcdpijr" "onzjrteqmu" "pquewclxuy" "vpmpffdoqz" "eouliovvra" "vxcbagyymm" "jekkafodvk" "ypekeuutef" "dlbqcynhrn" "erxulvebrj" "qwxrsgafzy" "dlsjwmqzhx" "exvhmqxvvp"
(integer) 100
```

Benchmark 
```
memtier_benchmark -S /tmp/2.socket --pipeline 10 --command="LRANGE list:100 0 -1"  --hide-histogram --test-time 60
```

#### results

Baseline
```
root@hpe10:~/redis# taskset -c 2-5 memtier_benchmark -S /tmp/2.socket --pipeline 10 --command="LRANGE list:100 0 -1"  --hide-histogram --test-time 60
Writing results to stdout
[RUN #1] Preparing benchmark client...
[RUN #1] Launching threads now...
[RUN #1 100%,  60 secs]  0 threads:    10140346 ops,  168711 (avg:  168966) ops/sec, 281.73MB/sec (avg: 282.15MB/sec), 11.74 (avg: 11.73) msec latency

4         Threads
50        Connections per thread
60        Seconds


ALL STATS
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Lranges    168969.88        11.72828        11.90300        17.02300        21.50300    288931.90 
Totals     168969.88        11.72828        11.90300        17.02300        21.50300    577863.80 
```

Change 1:
```
root@hpe10:~/redis# taskset -c 2-5 memtier_benchmark -S /tmp/1.socket --pipeline 10 --command="LRANGE list:100 0 -1"  --hide-histogram --test-time 60
Writing results to stdout
[RUN #1] Preparing benchmark client...
[RUN #1] Launching threads now...
[RUN #1 100%,  60 secs]  0 threads:    10456617 ops,  174327 (avg:  174243) ops/sec, 291.11MB/sec (avg: 290.97MB/sec), 11.36 (avg: 11.37) msec latency

4         Threads
50        Connections per thread
60        Seconds


ALL STATS
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Lranges    174239.75        11.36966        11.51900        16.51100        21.11900    297943.17 
Totals     174239.75        11.36966        11.51900        16.51100        21.11900    595886.33 

```

Change 3:

```
root@hpe10:~/redis# taskset -c 2-5 memtier_benchmark -S /tmp/1.socket --pipeline 10 --command="LRANGE list:100 0 -1"  --hide-histogram --test-time 60
Writing results to stdout
[RUN #1] Preparing benchmark client...
[RUN #1] Launching threads now...
[RUN #1 100%,  60 secs]  0 threads:    11386130 ops,  189935 (avg:  189733) ops/sec, 317.17MB/sec (avg: 316.83MB/sec), 10.42 (avg: 10.43) msec latency

4         Threads
50        Connections per thread
60        Seconds


ALL STATS
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Lranges    189726.81        10.43341        10.62300        15.16700        20.22300    324425.43 
Totals     189726.81        10.43341        10.62300        15.16700        20.22300    648850.86 

```


Change 4, commit 9a63d4d6a9fa946505e31ecce4c7796845fc022c:

```
root@hpe10:~/redis# taskset -c 2-5 memtier_benchmark -S /tmp/1.socket --pipeline 10 --command="LRANGE list:100 0 -1"  --hide-histogram --test-time 60
Writing results to stdout
[RUN #1] Preparing benchmark client...
[RUN #1] Launching threads now...
[RUN #1 100%,  60 secs]  0 threads:    11940572 ops,  199017 (avg:  198971) ops/sec, 332.34MB/sec (avg: 332.26MB/sec),  9.94 (avg:  9.94) msec latency

4         Threads
50        Connections per thread
60        Seconds


ALL STATS
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Lranges    198967.57         9.94374        10.11100        14.59100        19.58300    340226.77 
Totals     198967.57         9.94374        10.11100        14.59100        19.58300    680453.54 
```

Change 5, commit b544ade67628a1feaf714d6cfd114930e0c7670b:

```
root@hpe10:~/redis# taskset -c 2-5 memtier_benchmark -S /tmp/1.socket --pipeline 10 --command="LRANGE list:100 0 -1"  --hide-histogram --test-time 60
Writing results to stdout
[RUN #1] Preparing benchmark client...
[RUN #1] Launching threads now...
[RUN #1 100%,  60 secs]  0 threads:    11848352 ops,  197298 (avg:  197436) ops/sec, 329.47MB/sec (avg: 329.70MB/sec), 10.03 (avg: 10.02) msec latency

4         Threads
50        Connections per thread
60        Seconds


ALL STATS
==================================================================================================
Type         Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
--------------------------------------------------------------------------------------------------
Lranges    197427.70        10.02192        10.17500        14.71900        19.71100    337593.65 
Totals     197427.70        10.02192        10.17500        14.71900        19.71100    675187.31 

```