@fcostaoliveira 2024-11-14
### CE Performance Automation : step 1 of 2 (build) STARTING...

This comment was automatically generated given a benchmark was triggered.
Started building at 2024-11-14 09:59:36.726351
You can check each build/benchmark progress in grafana:
   - git hash: 9a63d4d6a9fa946505e31ecce4c7796845fc022c
   - git branch: filipecosta90:optimize.listpack
   - commit date and time: n/a
   - commit summary: n/a
   - test filters:
       - command priority lower limit: 0
       - command priority upper limit: 10000
       - test name regex: .*
       - command group regex: .*