@fcostaoliveira 2024-11-19
### Automated performance analysis summary

This comment was automatically generated given there is performance data available.

Using platform named: intel64-ubuntu22.04-redis-clx1 to do the comparison.

In summary:
- Detected a total of 1 stable tests between versions.

You can check a comparison in detail via the [grafana link](https://benchmarksredisio.grafana.net/d/1fWbtb7nz/experimental-oss-spec-benchmarks/?var-branch=unstable&var-branch=filipecosta90:optimize.listpack)

### Comparison between unstable and filipecosta90:optimize.listpack.

Time Period from 5 months ago. (environment used: oss-standalone)

<details>
  <summary>Full Results table:</summary>

|                                                                                                                      Test Case                                                                                                                      |Baseline redis/redis unstable (median obs. +- std.dev)|Comparison redis/redis filipecosta90:optimize.listpack (median obs. +- std.dev)|% change (higher-better)|  Note   |
|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|-----------------------------------------------------:|------------------------------------------------------------------------------:|------------------------|---------|
|[memtier_benchmark-10Kkeys-load-hash-50-fields-with-10000B-values](https://github.com/redis/redis-benchmarks-specification/blob/main/redis_benchmarks_specification/test-suites/memtier_benchmark-10Kkeys-load-hash-50-fields-with-10000B-values.yml)|                                                  1906|                                                                           1925|1.0%                    |No Change|

</details>